Initial commit
This more or less works to boot an iso image, at least on the happy path,
where everything matches expectations.

The big caveat is that the booted OS will need to know where to look for the
image, and that is not working / needs help.  (I had hoped the RamDisk
Protocol would register the image as a NVDIMM disk, but I'm not seeing it)